make it easier to be used as a python package
