+++ MrBam/aggregate.py
